requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile '.\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
colorama==0.4.6
    # via
    #   loguru
    #   tqdm
cramjam==2.11.0
    # via fastparquet
distro==1.9.0
    # via openai
fastparquet==2024.11.0
    # via -r .\requirements.in
filelock==3.18.0
    # via huggingface-hub
fsspec==2025.7.0
    # via
    #   fastparquet
    #   huggingface-hub
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
huggingface-hub==0.34.3
    # via -r .\requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via scikit-learn
loguru==0.7.3
    # via -r .\requirements.in
numpy==2.3.2
    # via
    #   -r .\requirements.in
    #   fastparquet
    #   pandas
    #   scikit-learn
    #   scipy
openai==1.97.1
    # via -r .\requirements.in
packaging==25.0
    # via
    #   fastparquet
    #   huggingface-hub
pandas==2.3.1
    # via
    #   -r .\requirements.in
    #   fastparquet
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r .\requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via huggingface-hub
requests==2.32.4
    # via
    #   -r .\requirements.in
    #   huggingface-hub
scikit-learn==1.7.1
    # via -r .\requirements.in
scipy==1.16.1
    # via scikit-learn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via
    #   -r .\requirements.in
    #   huggingface-hub
    #   openai
typing-extensions==4.14.1
    # via
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
win32-setctime==1.2.0
    # via loguru
